Crate kannader_config_macros

Macros that help define communication between a wasm blob and its rust host.

All the types are being exchanged bincode-encoded. If multiple parameters are to be taken, they are exchanged as a bincode-encoded tuple.

&mut references taken as arguments are taken as though they were by value, and then returned as supplementary arguments in a tuple.

Note: this crate’s guest-side implementation is very heavily tied to the kannader_config crate’s implementation. This is on purpose and the two crates should be used together. They are split only for technical reasons.

Macros

client_config_implement_guest_server

client_config_implement_host_client

client_config_implement_trait

implement_guest

implement_host

queue_config_implement_guest_server

queue_config_implement_host_client

queue_config_implement_trait

server_config_implement_guest_server

server_config_implement_host_client

server_config_implement_trait

tracing_implement_guest_client

tracing_implement_host_server

tracing_implement_trait